Move scan algo detail to parallel_backend_sycl.h
